moved shoot message details to own component
